backend/requirements.txt
aiocache==0.12.2
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
black==23.11.0
Brotli==1.1.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
Deprecated==1.2.14
dnspython==2.4.2
fastapi==0.104.1
firebase-admin==6.2.0
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.4
importlib-resources==6.1.1
limits==3.6.0
motor==3.3.2
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
slowapi==0.1.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.2
